+++ dchars/languages/hbo/dstring.py
@@ -48,6 +48,7 @@ from dchars.utilities.lstringtools import number_of_occurences
 
 # known transliterations :
 import dchars.languages.hbo.transliterations.basic.basic as basictrans
+import dchars.languages.hbo.transliterations.basic.ucombinations as basictrans_ucombinations
 
 ################################################################################
 class DStringHBO(DStringMotherClass):
@@ -77,6 +78,10 @@ class DStringHBO(DStringMotherClass):
     trans__init_from_transliteration = {
           "basic" : basictrans.dstring__init_from_translit_str,
           }
+
+    trans__get_transl_ucombinations = {
+        "basic"   : basictrans_ucombinations.get_usefull_combinations,
+        }
 
 
     #///////////////////////////////////////////////////////////////////////////
@@ -117,6 +122,24 @@ class DStringHBO(DStringMotherClass):
                 self.append( dchar )
         
         return self
+
+    #///////////////////////////////////////////////////////////////////////////
+    def get_usefull_transl_combinations(self):
+        """
+                DStringHBO.get_usefull_transl_combinations
+
+                Return a (str)string with all the usefull combinations of TRANSLITTERATED
+                characters, i.e. only the 'interesting' characters (not punctuation if
+                 it's too simple by example). 
+
+                NB : this function has nothing to do with linguistic or a strict
+                     approach of the language. This function allows only to get the
+                     most common and/or usefull characters of the writing system.
+
+                NB : function required by the dchars-fe project.
+        """
+        res = DStringHBO.trans__get_transl_ucombinations[self.transliteration_method]()
+        return res
 
     #///////////////////////////////////////////////////////////////////////////
     def get_transliteration(self):